test: moving fixed point test to proper folder
