+++ test/src/base/CMakeLists.txt
@@ -27,3 +27,11 @@ target_link_libraries(mp_utils_test
     mp_utils
     blob
     )
+
+
+addtest(fixed_point_test
+    fixed_point_test.cpp
+    )
+target_link_libraries(fixed_point_test
+    outcome
+)
